Allow -u|--unsorted and -a|--accelerated in any order in new()
The code for List::Compare::new() only detected correctly the unsorted and
accelerated options if they were given in this exact order.

This not-intuitive behaviour was described in a documentation note in the POD.

Moreover the code to detect -u|--unsorted was repeated in three locations:
List::Compare::new(), List::Compare::Base::_Auxiliary::_alt_construct_tester() and
List::Compare::Base::_Auxiliary::_alt_construct_tester_3().

Factored the code to detect -u|--unsorted and/or -a|--accelerated in any order in
List::Compare::Base::_Auxiliary::_parse_optios(), and called this new function from the
three places where the parsing was being done.

Changed t/04_oo_lists_dual_acc_unsorted.t to try all possible combinations of short/long
options, in any order, with presence/absence of each option and check the result of the
object construction to make sure the options were being take into account.

Removed the restriction of the options mandatory order from the POD.

Test all combinations of constructor with list and hash ref arguments

Add tests to call constructor with all option combinations in a hash-ref argument.

Fix indentation

Left "wrong" indentation in the previuous commit to better show the differences added.

Unrolled loops in test for all options combinations

Replaced for-loops that were testing all the combinations of options
-u, --unsorted, -a, --accelerated by the repeated sequence of test code.

Fix bug in tests for --unsorted: order of keys(%hash) is not predictable

The test for --unsorted assumed that the order of the keys of a hash
was predictable. The order changes with the Perl interpreter version.

Change the test to use keys(%expected) as the list to compare against,
which is the same as the module is using.